# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Recall CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, awaiting_return, at_risk]
        - name: barberId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/appointments:
    get:
      operationId: getClientAppointments
      tags: [clients]
      summary: Get appointment history for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"

  # ─── Appointments ────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons
      responses:
        "200":
          description: Coupon list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /reports/overview:
    get:
      operationId: getReportsOverview
      tags: [reports]
      summary: Get aggregated report metrics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: month
      responses:
        "200":
          description: Report overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportOverview"
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Daily and monthly revenue series
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /reports/clients:
    get:
      operationId: getClientReport
      tags: [reports]
      summary: New vs returning clients over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Client data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientDataPoint"

  # ─── AI Insights ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, nome, telefone, status, ativo, totalVisitas, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        email:
          type: ["string", "null"]
        dataNascimento:
          type: ["string", "null"]
          format: date
        observacoes:
          type: ["string", "null"]
        # Derivado do histórico de atendimentos (skill client-recall-logic):
        # Ativo / Aguardando Retorno / Em Risco. Somente leitura.
        status:
          type: string
          enum: [active, awaiting_return, at_risk]
        ativo:
          type: boolean
        totalVisitas:
          type: integer
        ultimoAtendimento:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [nome, telefone]
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
          format: date
        observacoes:
          type: string
    ClientUpdate:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
        dataNascimento:
          type: string
          format: date
        observacoes:
          type: string
        ativo:
          type: boolean

    # ─── Barber ─────────────────────────────────────────────────────
...
    Appointment:
      type: object
      required: [id, clienteId, valor, valorFinal, data, createdAt]
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        clienteNome:
          type: ["string", "null"]
        barbeiroId:
          type: ["integer", "null"]
        barbeiroNome:
          type: ["string", "null"]
        servicoId:
          type: ["integer", "null"]
        servicoNome:
          type: ["string", "null"]
        valor:
          type: number
        desconto:
          type: number
        valorFinal:
          type: number
        data:
          type: string
          format: date-time
        observacoes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Coupon:
      type: object
      required: [id, codigo, tipo, valor, ativo, usoAtual]
      properties:
        id:
          type: integer
        codigo:
          type: string
        tipo:
          type: string
          enum: [percent, fixed]
        valor:
          type: number
        validade:
          type: ["string", "null"]
          format: date
        ativo:
          type: boolean
        usoMaximo:
          type: ["integer", "null"]
        usoAtual:
          type: integer
    CouponInput:
      type: object
      required: [tipo, valor]
      properties:
        codigo:
          type: string
        tipo:
          type: string
          enum: [percent, fixed]
        valor:
          type: number
        validade:
          type: string
          format: date
        usoMaximo:
          type: integer
    CouponUpdate:
      type: object
      properties:
        codigo:
          type: string
        tipo:
          type: string
          enum: [percent, fixed]
        valor:
          type: number
        validade:
          type: string
          format: date
        ativo:
          type: boolean
        usoMaximo:
          type: integer

    # ─── Dashboard ──────────────────────────────────────────────────
...
    ReportOverview:
      type: object
      required: [receitaDiaria, receitaMensal, clientesNovos, clientesRecorrentes, tempoMedioRetorno, ticketMedio, cuponsUsados, campanhasEnviadas, taxaAbertura, taxaRetorno]
      properties:
        receitaDiaria:
          type: number
        receitaMensal:
          type: number
        clientesNovos:
          type: integer
        clientesRecorrentes:
          type: integer
        tempoMedioRetorno:
          type: number
        ticketMedio:
          type: number
        cuponsUsados:
          type: integer
        campanhasEnviadas:
          type: integer
        taxaAbertura:
          type: number
        taxaRetorno:
          type: number
    RevenueDataPoint:
      type: object
      required: [date, revenue, appointments]
      properties:
        date:
          type: string
        revenue:
          type: number
        appointments:
          type: integer
    ClientDataPoint:
      type: object
      required: [date, novos, recorrentes]
      properties:
        date:
          type: string
        novos:
          type: integer
        recorrentes:
          type: integer

    # ─── Insights ───────────────────────────────────────────────────
...